change type of cursor position from int to double

+++ include/zen/scene/screen-layout.h
@@ -21,7 +21,8 @@ struct zn_screen_layout {
 // Returns the screen at the given x, y of the coordinates of screen layout.
 // And sets the screen-local coordinates to dst_{x,y}
 struct zn_screen* zn_screen_layout_get_closest_screen(
-    struct zn_screen_layout* self, int x, int y, int* dst_x, int* dst_y);
+    struct zn_screen_layout* self, double x, double y, double* dst_x,
+    double* dst_y);
 
 void zn_screen_layout_add(
     struct zn_screen_layout* self, struct zn_screen* screen);
